requirements.txt
python-telegram-bot==20.0a6
requests>=2.28.0
beautifulsoup4>=4.11.1
python-dotenv>=0.21.0